feat: Complete Phase 5 - CPU extensions and addressing modes
- Add full flag support (ZeroFlag, CarryFlag, OverflowFlag, SignFlag) in all arithmetic instructions
- Fix CarryFlag logic in AddInstruction (use ulong for unsigned overflow detection)
- Add CarryFlag to IncInstruction and DecInstruction
- Implement AddressingMode support (Immediate, Direct, Indirect, Relative) in:
  - LoadInstruction, StoreInstruction
  - JumpInstruction, JumpIfZeroInstruction, JumpIfNotZeroInstruction
  - CallInstruction
- Add FlagTests.cs with 15 tests for all flag combinations
- Add AddressingModeTests.cs with 8 tests for Indirect and Relative modes
- Fix test assertions for correct flag behavior
- All 59 tests pass

Part of modernization plan (plans/002-cpu-modernization.md)

+++ src/CpuEmulator/Execution/Instructions/AddInstruction.cs
@@ -10,6 +10,13 @@ namespace CpuEmulator.Execution.Instructions;
 public class AddInstruction : IInstruction
 {
     /// <inheritdoc />
+    /// <summary>
+    /// Wykonuje instrukcję Add - dodaje wartości dwóch rejestrów i ustawia flagi.
+    /// </summary>
+    /// <param name="state">Aktualny stan CPU.</param>
+    /// <param name="instruction">Instrukcja do wykonania.</param>
+    /// <returns>Nowy stan CPU z zaktualizowanym rejestrem i flagami.</returns>
+    /// <exception cref="InvalidOperandException">Rzucane, gdy indeks rejestru źródłowego lub docelowego jest nieprawidłowy.</exception>
     public CpuState Execute(CpuState state, Instruction instruction)
     {
         if (instruction.Operand1 < 0 || instruction.Operand1 >= state.Registers.Count)
@@ -32,7 +39,7 @@ public class AddInstruction : IInstruction
         var newFlags = newState.Flags
             .WithZeroFlag(result == 0)
             .WithSignFlag(result < 0)
-            .WithCarryFlag((uint)value1 + (uint)value2 > uint.MaxValue)
+            .WithCarryFlag((ulong)(uint)value1 + (ulong)(uint)value2 > uint.MaxValue)
             .WithOverflowFlag(IsOverflow(value1, value2, result));
 
         return newState.WithFlags(newFlags);
